refactor(ale_claw): Tier 4b — group canonical modules into canonical/ subpackage
Moves canonical, canonical_types, canonical_sanitize into harness/canonical/ via
git mv (names kept). The trio imports only each other (siblings → unchanged);
canonical_types has zero external importers. External `.canonical*` importers
(agent_loop, memory_flush, replay, compaction, __init__, 5 tests) updated to
.canonical.canonical*. GOTCHA fixed: canonical.py has two LAZY (in-method)
`from .model_config import resolve_model` calls that the top-level-import scan
missed — they resolved inside the subpackage and broke get_transcript_policy
(14 test failures); changed to `..model_config`. Pure move; full suite green
(928 passed, 3 skipped).
