+++ ale_run/agents/ale_claw/harness/compaction.py
@@ -13,7 +13,7 @@ import json
 from dataclasses import dataclass
 from typing import Any
 
-from .canonical_sanitize import _SKIP_SYNTHESIS_STOP_REASONS
+from .canonical.canonical_sanitize import _SKIP_SYNTHESIS_STOP_REASONS
 from .helper_runtime import call_helper_model
 from .model_config import ResolvedModel, resolve_model
 from .token_estimation import (
