+++ tests/ale_claw/test_sanitize_items.py
@@ -11,7 +11,7 @@ Covers:
 
 import json
 
-from ale_run.agents.ale_claw.harness.canonical import (
+from ale_run.agents.ale_claw.harness.canonical.canonical import (
     CanonicalMessage,
     CompactionSummaryBlock,
     ComputerCallBlock,
